+++ Source/WinCopies.Util.Shared/Delegates/DelegateTypes.cs
@@ -121,6 +121,10 @@ namespace WinCopies
 #endif
         T>(object obj);
 
+    public delegate void EventHandler<T>(T sender, EventArgs e);
+
+    public delegate void EventHandler<TSender, TEventArgs>(TSender sender, TEventArgs e);
+
     public delegate void ActionParams(params object[] args);
 
     public delegate void ActionParams<in T>(params T[] args);
@@ -133,56 +137,72 @@ namespace WinCopies
 
     public delegate object FuncParams(params object[] args);
 
-    public delegate TResult FuncParams<in TParams, out TResult>(params TParams[] args);
+    public delegate TOut FuncParams<in TParams, out TOut>(params TParams[] args);
 
-    public delegate TResult FuncOut<in T1, T2, out TResult>(T1 p1, out T2 p2);
-    public delegate TResult FuncOut<in T1, in T2, T3, out TResult>(T1 p1, T2 p2, out T3 p3);
-    public delegate TResult FuncOut<in T1, in T2, in T3, T4, out TResult>(T1 p1, T2 p2, T3 p3, out T4 p4);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, T5, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, out T5 p5);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, T6, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, out T6 p6);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, T7, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, out T7 p7);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, T8, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, out T8 p8);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, T9, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, out T9 p9);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, T10, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, out T10 p10);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, T11, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, out T11 p11);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, T12, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, out T12 p12);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, T13, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, out T13 p13);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, in T13, T14, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, T13 p13, out T14 p14);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, in T13, in T14, T15, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, T13 p13, T14 p14, out T15 p15);
-    public delegate TResult FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, in T13, in T14, in T15, T16, out TResult>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, T13 p13, T14 p14, T15 p15, out T16 p16);
+    public delegate TOut FuncIn<T1, out TOut>(in T1 p1);
+    public delegate TOut FuncIn<T1, T2, out TOut>(in T1 p1, in T2 p2);
+    public delegate TOut FuncIn<T1, T2, T3, out TOut>(in T1 p1, in T2 p2, in T3 p3);
+    public delegate TOut FuncIn<T1, T2, T3, T4, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, in T15 p15);
+    public delegate TOut FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, in T15 p15, in T16 p16);
 
-    public delegate TResult FuncIn<T1, out TResult>(in T1 p1);
-    public delegate TResult FuncIn<T1, T2, out TResult>(in T1 p1, in T2 p2);
-    public delegate TResult FuncIn<T1, T2, T3, out TResult>(in T1 p1, in T2 p2, in T3 p3);
-    public delegate TResult FuncIn<T1, T2, T3, T4, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, in T15 p15);
-    public delegate TResult FuncIn<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, in T15 p15, in T16 p16);
+    public delegate TOut FuncIn2<in T1, T2, out TOut>(T1 value1, in T2 value2);
+    public delegate TOut FuncIn2<in T1, T2, T3, out TOut>(T1 value1, in T2 value2, in T3 value3);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9, in T10 value10);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9, in T10 value10, in T11 value11);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9, in T10 value10, in T11 value11, in T12 value12);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9, in T10 value10, in T11 value11, in T12 value12, in T13 value13);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9, in T10 value10, in T11 value11, in T12 value12, in T13 value13, in T14 value14);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9, in T10 value10, in T11 value11, in T12 value12, in T13 value13, in T14 value14, in T15 value15);
+    public delegate TOut FuncIn2<in T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, out TOut>(T1 value1, in T2 value2, in T3 value3, in T4 value4, in T5 value5, in T6 value6, in T7 value7, in T8 value8, in T9 value9, in T10 value10, in T11 value11, in T12 value12, in T13 value13, in T14 value14, in T15 value15, in T16 value16);
 
-    public delegate TResult FuncInOut<T1, T2, out TResult>(in T1 p1, out T2 p2);
-    public delegate TResult FuncInOut<T1, T2, T3, out TResult>(in T1 p1, in T2 p2, out T3 p3);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, out TResult>(in T1 p1, in T2 p2, in T3 p3, out T4 p4);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, out T5 p5);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, out T6 p6);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, out T7 p7);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, out T8 p8);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, out T9 p9);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, out T10 p10);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, out T11 p11);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, out T12 p12);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, out T13 p13);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, out T14 p14);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, out T15 p15);
-    public delegate TResult FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, out TResult>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, in T15 p15, out T16 p16);
+    public delegate TOut FuncOut<in T1, T2, out TOut>(T1 p1, out T2 p2);
+    public delegate TOut FuncOut<in T1, in T2, T3, out TOut>(T1 p1, T2 p2, out T3 p3);
+    public delegate TOut FuncOut<in T1, in T2, in T3, T4, out TOut>(T1 p1, T2 p2, T3 p3, out T4 p4);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, T5, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, out T5 p5);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, T6, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, out T6 p6);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, T7, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, out T7 p7);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, T8, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, out T8 p8);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, T9, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, out T9 p9);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, T10, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, out T10 p10);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, T11, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, out T11 p11);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, T12, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, out T12 p12);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, T13, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, out T13 p13);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, in T13, T14, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, T13 p13, out T14 p14);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, in T13, in T14, T15, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, T13 p13, T14 p14, out T15 p15);
+    public delegate TOut FuncOut<in T1, in T2, in T3, in T4, in T5, in T6, in T7, in T8, in T9, in T10, in T11, in T12, in T13, in T14, in T15, T16, out TOut>(T1 p1, T2 p2, T3 p3, T4 p4, T5 p5, T6 p6, T7 p7, T8 p8, T9 p9, T10 p10, T11 p11, T12 p12, T13 p13, T14 p14, T15 p15, out T16 p16);
+
+    public delegate TOut FuncInOut<T1, T2, out TOut>(in T1 p1, out T2 p2);
+    public delegate TOut FuncInOut<T1, T2, T3, out TOut>(in T1 p1, in T2 p2, out T3 p3);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, out TOut>(in T1 p1, in T2 p2, in T3 p3, out T4 p4);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, out T5 p5);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, out T6 p6);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, out T7 p7);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, out T8 p8);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, out T9 p9);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, out T10 p10);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, out T11 p11);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, out T12 p12);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, out T13 p13);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, out T14 p14);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, out T15 p15);
+    public delegate TOut FuncInOut<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, out TOut>(in T1 p1, in T2 p2, in T3 p3, in T4 p4, in T5 p5, in T6 p6, in T7 p7, in T8 p8, in T9 p9, in T10 p10, in T11 p11, in T12 p12, in T13 p13, in T14 p14, in T15 p15, out T16 p16);
 
     public delegate void ActionIn<T>(in T parameter);
     public delegate void ActionIn<T1, T2>(in T1 p1, in T2 p2);